+++ src/Widgets/BasicWidgets.hpp
@@ -16,6 +16,6 @@ namespace osc
 {
 	void DrawComponentHoverTooltip(OpenSim::Component const&);
 	void DrawSelectOwnerMenu(osc::VirtualModelStatePair&, OpenSim::Component const&);
-	void DrawRequestOutputsMenu(osc::MainUIStateAPI&, OpenSim::Component const&);
+	void DrawWatchOutputMenu(osc::MainUIStateAPI&, OpenSim::Component const&);
 	void DrawSimulationParams(osc::ParamBlock const&);
 }